- added valid cd check for SDL again (please don't remove it again, a   ctp2 executable without the check would be against the eula) - readded linux specific code in c3files - fixed a warning in spritefile.cpp - linux compile fixes in gfx/tilesys - compilation fixes in libs/anet - StupidPlaySound needs linux implementation
API changes:
- added aui_NativeSurface typedef (either aui_DirectSurface or aui_SDLSurface)
- aui_NativeSurface->Blank and BlankRGB will fill a surface with the color
  given; i.e. all pixels will be overwritten. TODO: bpp check


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@979 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gfx/spritesys/spritefile.cpp
@@ -2243,7 +2243,7 @@ long SpriteFile::GetFilePos(void)
 	Assert(err == 0);
 
 #ifdef WIN32
-	return m_filePos;
+	return static_cast<long>(m_filePos);
 #elif defined(LINUX)
 	return m_filePos.__pos;
 #endif
